Refactor dashboard header
